Move isAuthorized to component root
